+++ templates/blocks/basic_rebates-who_to_contact.php
@@ -6,10 +6,36 @@
  *
  * @package Bcgov\Plugin\BasicBlocks
  */
-?>
+require WP_PLUGIN_DIR . '/bcgov-plugin-cleanbc/templates/config/template-vars.php';
 
-<?php if ( $is_gb_editor ) { ?>
-	<p>*In Editor*</p>
-<?php } ?>
+/**
+ * ACF Fields
+ */
+$who_to_contact_section = get_field( 'who_to_contact_section', $post->ID );
+$content_blocks         = $who_to_contact_section['content_blocks'];
 
- <p>TODO: Rebates Page - who_to_contact Content</p>
+if ( $is_gb_editor ) : ?>
+	<h2>Who to Contact</h2>
+	<p><b>Note to editors:</b> This content can be found in the <b>Who to Contact</b> tab of the <b>Template Rebate</b> section in the post itself.</p>
+<?php elseif ( $content_blocks ) : ?>
+	<section id="who-to-contact" class="block block--whotocontact whotocontact-block">
+		<div class="inner">
+			<h2>Who to Contact</h2>
+			<div class="content-blocks">
+				<div class="inner">
+					<?php
+                    foreach ( $content_blocks as $key => $content_block ) :
+						$content_block_classes   = [
+							'content-block ',
+						];
+						$content_block_layout    = $content_block['acf_fc_layout'];
+						$content_block_classes[] = $content_block_layout . ' ';
+
+						include WP_PLUGIN_DIR . '/bcgov-plugin-cleanbc/templates/blocks/partials/content-blocks.php';
+					endforeach;
+                    ?>
+				</div>
+			</div>
+		</div>
+	</section>
+<?php endif; ?>